schema-profiler-agent/pyproject.toml
[tool.poetry]
name = "schema-profiler-agent"
version = "0.1.0"
description = "An agent for profiling datasets and generating schema configurations."
authors = ["Your Name <you@example.com>"]
license = "Apache-2.0"
readme = "README.md"
packages = [{include = "schema_profiler"}]

[tool.poetry.dependencies]
python = "^3.13"
google-adk = { version = "^1.2.0", extras = ["mcp"] }
pandas = "^2.2.0"
tableauhyperapi = ">=0.0.18312"
pyyaml = "^6.0"
watchdog = "^3.0.0" # For monitoring file system events
fastapi = "^0.115.0"
uvicorn = "^0.34.0"
google-cloud-aiplatform = { extras = ["adk", "agent-engines"], version = "^1.93.0" }
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"
google-genai = "^1.16.1"
httpx = "^0.28.0"
aiohttp = "^3.10.0"
sse-starlette = "^2.3.0"
pyjwt = "^2.10.0"
jwcrypto = "^1.5.0"
openpyxl = "^3.1.0"
tenacity = "^8.2.0"
a2a-sdk = "^0.2.12"
litellm = "^1.0.0"
google-generativeai = "^0.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-mock = "^3.0"
pytest-cov = "^3.0"
psutil = "^5.9.0"
requests = "^2.31.0"

[tool.poetry.group.deployment]
optional = true

[tool.poetry.group.deployment.dependencies]
absl-py = "^2.2.1"
cloudpickle = "^3.1.1"
flake8-pyproject = "^1.2.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
